Relax product and quotient constraints.
Since interval multiplication and division is not defined when c == 0 or
d == 0, we have to determine some default action. We can't narrow the
variable, but it's not always invalid. If the value is a singleton and
is zero, then we know we don't have a solution. If the interval is
wider, we may have a solution after narrowing it down.