Remove symmetric constraints.
Difference can be expressed as sum, and quotient can be expressed as
product with the additional constraint that the denominator is not equal
to zero in any solution. For now we create an extra hidden variable and
add an inequality constraint. If we add NotEqualToConstant constraints
in the future we can replace it with that (not sure if it is more
elegant that way).