Implement better pruning.
Domain iterators were previously iterating all singleton domain subsets,
which was very expensive for large domains.

Domain iterators may now return (disjoint) chunks of their domain to
test to see if any of the subvalues are promising. For integer domains,
this allows huge ranges of numbers to be discarded immediately.